refactor(scraping): Extract person workflow
Move scrape_person, get_my_profile, get_sidebar_profiles and
search_people onto a PersonScraper, and the profile URN and display-name
reads onto a ProfilePageReader. The facade keeps all four entry points as
delegates with unchanged signatures.

The sidebar workflow's English strings move into an explicit per-locale
table in text.py, which person.py templates its extraction program from,
so the one text exception stays written down once.

The URN read borrows the facade's top-card read through an injected
callable, because that read stays on the facade until the message sender
owns it.
