+++ linkedin_mcp_server/scraping/extractor.py
@@ -69,6 +69,8 @@ from linkedin_mcp_server.scraping.job_policy import (
     same_job_search,
 )
 from linkedin_mcp_server.scraping.navigation import PageNavigator
+from linkedin_mcp_server.scraping.person import NAV_DELAY, PersonScraper
+from linkedin_mcp_server.scraping.profile_page import ProfilePageReader
 from linkedin_mcp_server.scraping.session import ScrapingSession
 from linkedin_mcp_server.scraping.link_metadata import (
     Reference,
@@ -79,7 +81,6 @@ from linkedin_mcp_server.scraping.search_urls import (
     build_company_search_url,
     build_content_search_url,
     build_job_search_url,
-    build_people_search_url,
 )
 from linkedin_mcp_server.scraping.text import (
     filter_linkedin_noise_lines,
@@ -88,15 +89,12 @@ from linkedin_mcp_server.scraping.text import (
     truncate_linkedin_noise,
 )
 
-from .fields import COMPANY_SECTIONS, PERSON_SECTIONS
+from .fields import COMPANY_SECTIONS
 
 if TYPE_CHECKING:
     from linkedin_mcp_server.callbacks import ProgressCallback
 
 logger = logging.getLogger(__name__)
-
-# Pacing between page navigations
-_NAV_DELAY = 2.0
 
 # The id is the trailing run of digits, and LinkedIn serves the same job under
 # both `/jobs/view/1967281839/` and `/jobs/view/<title>-at-<company>-1967281839/`.
@@ -1376,6 +1374,15 @@ class LinkedInExtractor:
         self._content = PageContentReader(self._session)
         self._capture = SectionCapture(self._session, self._navigator, self._content)
         self._feed = FeedScraper(self._session, self._navigator, self._content)
+        # Late-bound on purpose: the top-card read the URN comes from still
+        # lives here until the message sender owns it, and a bound method
+        # captured now would not see a replacement installed on this instance.
+        self._profile_page = ProfilePageReader(
+            self._session, lambda: self._read_profile_message_target()
+        )
+        self._person = PersonScraper(
+            self._session, self._navigator, self._capture, self._profile_page
+        )
         self._page = page
         # What the sidebar scroll spent on the page being read, so that a
         # multi-page search charges its scroll budget for scrolling alone.
@@ -1689,145 +1696,15 @@ class LinkedInExtractor:
         main_profile_already_loaded: bool = False,
         allow_self_alias: bool = False,
     ) -> dict[str, Any]:
-        """Scrape a person profile with configurable sections.
-
-        When ``main_profile_already_loaded`` is True and ``self._page`` is on
-        the exact profile root for ``username``, the ``main_profile`` section
-        is extracted from the current page without re-navigating. Falls back
-        to ``extract_page`` if the URL drifts or the reuse path returns the
-        soft-rate-limit sentinel (preserving the retry semantics of
-        ``extract_page``).
-
-        Returns:
-            {url, sections: {name: text}, profile_urn?: str}
-        """
-        requested = requested | {"main_profile"}
-        username = normalize_person_identifier(
-            username, allow_self_alias=allow_self_alias
+        """Scrape a person profile with configurable sections."""
+        return await self._person.scrape_person(
+            username,
+            requested,
+            callbacks,
+            max_scrolls,
+            main_profile_already_loaded=main_profile_already_loaded,
+            allow_self_alias=allow_self_alias,
         )
-        base_url = person_profile_url(username)
-        sections: dict[str, str] = {}
-        references: dict[str, list[Reference]] = {}
-        section_errors: dict[str, dict[str, Any]] = {}
-        profile_urn: str | None = None
-        rate_limited = False
-
-        requested_ordered = [
-            (name, suffix, is_overlay)
-            for name, (suffix, is_overlay) in PERSON_SECTIONS.items()
-            if name in requested
-        ]
-        total = len(requested_ordered)
-
-        if callbacks:
-            await callbacks.on_start("person profile", base_url)
-
-        try:
-            for i, (section_name, suffix, is_overlay) in enumerate(requested_ordered):
-                if i > 0:
-                    await asyncio.sleep(_NAV_DELAY)
-
-                url = base_url + suffix
-                try:
-                    can_reuse_main = (
-                        section_name == "main_profile"
-                        and main_profile_already_loaded
-                        and urlparse(self._page.url).path.rstrip("/")
-                        == urlparse(base_url).path.rstrip("/")
-                    )
-                    if can_reuse_main:
-                        extracted = await self._capture._extract_loaded_section(
-                            url,
-                            section_name=section_name,
-                            max_scrolls=max_scrolls,
-                        )
-                        if extracted.text == RATE_LIMITED_SECTION_TEXT:
-                            logger.info(
-                                "Reuse path soft-rate-limited; falling back "
-                                "to extract_page for retry parity"
-                            )
-                            extracted = await self.extract_page(
-                                url,
-                                section_name=section_name,
-                                max_scrolls=max_scrolls,
-                            )
-                    elif is_overlay:
-                        extracted = await self._capture._extract_overlay(
-                            url, section_name=section_name
-                        )
-                    else:
-                        extracted = await self.extract_page(
-                            url,
-                            section_name=section_name,
-                            max_scrolls=max_scrolls,
-                        )
-
-                    if extracted.text and extracted.text != RATE_LIMITED_SECTION_TEXT:
-                        sections[section_name] = extracted.text
-                        if extracted.references:
-                            references[section_name] = extracted.references
-                    elif extracted.text == RATE_LIMITED_SECTION_TEXT:
-                        section_errors[section_name] = rate_limited_section_error()
-                        # Stop rather than walk the remaining sections. Each one
-                        # is another navigation, and LinkedIn has just said it
-                        # wants fewer of them. Whatever was gathered before this
-                        # point is kept and returned.
-                        rate_limited = True
-                    elif extracted.error:
-                        section_errors[section_name] = extracted.error
-
-                    # Skipped once the section came back empty: there is no
-                    # content to read a URN from, and a failure here lands in
-                    # the handler below, which would overwrite the entry just
-                    # recorded with a generic diagnostic — losing the one
-                    # finding this section had.
-                    if (
-                        section_name == "main_profile"
-                        and profile_urn is None
-                        and not rate_limited
-                    ):
-                        profile_urn = await self._extract_profile_urn()
-                except LinkedInScraperException:
-                    raise
-                except Exception as e:
-                    logger.warning("Error scraping section %s: %s", section_name, e)
-                    section_errors[section_name] = build_issue_diagnostics(
-                        e,
-                        context="scrape_person",
-                        target_url=url,
-                        section_name=section_name,
-                    )
-
-                # "Scraped" = processed/attempted, not necessarily successful.
-                # Per-section failures are captured in section_errors.
-                if callbacks:
-                    percent = round((i + 1) / total * 95)
-                    await callbacks.on_progress(
-                        f"Scraped {section_name} ({i + 1}/{total})", percent
-                    )
-
-                if rate_limited:
-                    break
-        except LinkedInScraperException as e:
-            if callbacks:
-                await callbacks.on_error(e)
-            raise
-
-        result: dict[str, Any] = {
-            "url": f"{base_url}/",
-            "sections": sections,
-        }
-        if profile_urn:
-            result["profile_urn"] = profile_urn
-        if references:
-            result["references"] = references
-        if section_errors:
-            result["section_errors"] = section_errors
-
-        if callbacks:
-            await callbacks.on_complete("person profile", result)
-
-        return result
 
     async def get_my_profile(
         self,
@@ -1835,32 +1712,8 @@ class LinkedInExtractor:
         callbacks: ProgressCallback | None = None,
         max_scrolls: int | None = None,
     ) -> dict[str, Any]:
-        """Scrape the authenticated user's own LinkedIn profile.
-
-        Navigates to /in/me/ and resolves the redirect to obtain the real
-        username before scraping, so result["url"] reflects the actual profile
-        URL rather than /in/me/.
-
-        Returns:
-            {url, sections: {name: text}}
-        """
-        await self._navigator._navigate_to_page("https://www.linkedin.com/in/me/")
-        real_url = self._page.url  # post-redirect, e.g. /in/johndoe/
-        match = re.search(r"/in/([^/?#]+)", real_url)
-        username = match.group(1) if match else "me"
-        logger.debug("get_my_profile resolved username=%r from %s", username, real_url)
-
-        return await self.scrape_person(
-            username,
-            sections if sections is not None else {"main_profile"},
-            callbacks=callbacks,
-            max_scrolls=max_scrolls,
-            main_profile_already_loaded=True,
-            # The redirect is what resolves the alias. When it has not, this is
-            # still the tool the user asked for, so "me" stays usable here and
-            # nowhere else.
-            allow_self_alias=True,
-        )
+        """Scrape the authenticated user's own LinkedIn profile."""
+        return await self._person.get_my_profile(sections, callbacks, max_scrolls)
 
     async def _read_action_signals(self, username: str) -> ActionSignals:
         """Read locale-independent structural signals for a profile's
@@ -2272,187 +2125,9 @@ class LinkedInExtractor:
             profile=verified_text or page_text,
         )
 
-    async def _extract_profile_urn(self) -> str | None:
-        """Extract a profile URN only from one unambiguous top-card snapshot."""
-        resolution = await self._read_profile_message_target()
-        return resolution.target.profile_urn if resolution.target else None
-
     async def get_sidebar_profiles(self, username: str) -> dict[str, Any]:
-        """Extract profile links from sidebar sections on a LinkedIn profile page.
-
-        Scrapes "More profiles for you", "Explore premium profiles", and
-        "People you may know" sidebar sections. Follows each "Show all" link to
-        collect the full list; skips any section whose "Show all" URL contains or
-        redirects to /premium.
-
-        Returns:
-            Dict with url and sidebar_profiles mapping section key to list of
-            /in/username/ paths. Sections absent from the page are omitted.
-        """
-        username = normalize_person_identifier(username)
-        url = person_profile_url(username, "/")
-        await self._navigator._navigate_to_page(url)
-        await detect_rate_limit(self._page)
-
-        try:
-            await self._page.wait_for_selector("main", timeout=5000)
-        except PlaywrightTimeoutError:
-            logger.debug("No <main> element found on %s", url)
-
-        await handle_modal_close(self._page)
-
-        sidebar_data: dict[str, Any] = await self._page.evaluate(
-            """() => {
-                const SIDEBAR_SECTIONS = [
-                    "More profiles for you",
-                    "Explore premium profiles",
-                    "People you may know"
-                ];
-                const normalize = text => (text || '').replace(/\\s+/g, ' ').trim();
-                const slugify = text => text.toLowerCase().replace(/\\s+/g, '_');
-                const extractProfilePath = href => {
-                    if (!href) return null;
-                    const idx = href.indexOf('/in/');
-                    if (idx === -1) return null;
-                    const rest = href.slice(idx + 4);
-                    const end = rest.search(/[/?#]/);
-                    const username = end === -1 ? rest : rest.slice(0, end);
-                    return username ? '/in/' + username + '/' : null;
-                };
-
-                const sections = {};
-                const showAllUrls = {};
-
-                const headings = Array.from(document.querySelectorAll('h1, h2, h3'));
-                for (const heading of headings) {
-                    const headingText = normalize(
-                        heading.innerText || heading.textContent
-                    );
-                    if (!SIDEBAR_SECTIONS.includes(headingText)) continue;
-
-                    const sectionKey = slugify(headingText);
-
-                    // Walk up to find a section/aside container (max 5 levels)
-                    let container = heading.parentElement;
-                    let foundSection = false;
-                    for (let depth = 0; container && depth < 5; depth++) {
-                        const tag = container.tagName.toLowerCase();
-                        if (tag === 'section' || tag === 'aside') { foundSection = true; break; }
-                        container = container.parentElement;
-                    }
-                    if (!container || !foundSection) continue;
-
-                    // Collect /in/ profile links, deduplicated
-                    const seen = new Set();
-                    const profileLinks = [];
-                    for (const a of container.querySelectorAll('a[href*="/in/"]')) {
-                        const path = extractProfilePath(a.getAttribute('href'));
-                        if (path && !seen.has(path)) {
-                            seen.add(path);
-                            profileLinks.push(path);
-                        }
-                    }
-
-                    // Find "Show all" / "See all" anchor within container
-                    let showAll = null;
-                    for (const a of container.querySelectorAll('a')) {
-                        const text = normalize(
-                            a.innerText || a.textContent
-                        ).toLowerCase();
-                        if (text.startsWith('show all') || text.startsWith('see all')) {
-                            showAll = a.href || a.getAttribute('href');
-                            break;
-                        }
-                    }
-
-                    sections[sectionKey] = profileLinks;
-                    if (showAll) showAllUrls[sectionKey] = showAll;
-                }
-
-                return { sections, showAllUrls };
-            }"""
-        )
-
-        sidebar_profiles: dict[str, list[str]] = dict(sidebar_data.get("sections", {}))
-        show_all_urls: dict[str, str] = dict(sidebar_data.get("showAllUrls", {}))
-
-        first_show_all = True
-        for section_key, show_all_url in show_all_urls.items():
-            if "/premium" in show_all_url:
-                continue
-
-            if not first_show_all:
-                await asyncio.sleep(_NAV_DELAY)
-            first_show_all = False
-
-            try:
-                await self._navigator._navigate_to_page(show_all_url)
-            except LinkedInScraperException:
-                raise
-            except Exception:
-                logger.debug(
-                    "Failed to navigate to Show all for section %s: %s",
-                    section_key,
-                    show_all_url,
-                )
-                continue
-
-            if "/premium" in self._page.url:
-                logger.debug(
-                    "Show all for section %s redirected to premium, skipping",
-                    section_key,
-                )
-                continue
-
-            await detect_rate_limit(self._page)
-
-            try:
-                await self._page.wait_for_selector("main")
-            except PlaywrightTimeoutError:
-                logger.debug("No <main> on Show all page for section %s", section_key)
-
-            await handle_modal_close(self._page)
-
-            expanded_links: list[str] = await self._page.evaluate(
-                """() => {
-                    const extractProfilePath = href => {
-                        if (!href) return null;
-                        const idx = href.indexOf('/in/');
-                        if (idx === -1) return null;
-                        const rest = href.slice(idx + 4);
-                        const end = rest.search(/[/?#]/);
-                        const username = end === -1 ? rest : rest.slice(0, end);
-                        return username ? '/in/' + username + '/' : null;
-                    };
-                    const seen = new Set();
-                    const links = [];
-                    for (const a of document.querySelectorAll(
-                        'main a[href*="/in/"]'
-                    )) {
-                        const path = extractProfilePath(a.getAttribute('href'));
-                        if (path && !seen.has(path)) {
-                            seen.add(path);
-                            links.push(path);
-                        }
-                    }
-                    return links;
-                }"""
-            )
-
-            # Merge: sidebar links first, then show_all expansion, deduped
-            existing = sidebar_profiles.get(section_key, [])
-            seen_paths: set[str] = set(existing)
-            merged = list(existing)
-            for link in expanded_links:
-                if link not in seen_paths:
-                    seen_paths.add(link)
-                    merged.append(link)
-            sidebar_profiles[section_key] = merged
-
-        return {
-            "url": url,
-            "sidebar_profiles": sidebar_profiles,
-        }
+        """Extract profile links from sidebar sections on a profile page."""
+        return await self._person.get_sidebar_profiles(username)
 
     async def _read_profile_message_target(self) -> _ProfileMessageTargetResolution:
         """Resolve one recipient-specific top-card compose action after settling."""
@@ -2521,33 +2196,6 @@ class LinkedInExtractor:
         """Return an unambiguous recipient-specific top-card compose URL."""
         resolution = await self._read_profile_message_target()
         return resolution.target.compose_url if resolution.target else None
-
-    async def _read_profile_display_name(self) -> str | None:
-        """Read the visible profile name from the current person page."""
-        display_name = await self._page.evaluate(
-            """() => {
-                const heading = document.querySelector('main h1');
-                const normalize = value => (value || '').replace(/\\s+/g, ' ').trim();
-                if (heading) {
-                    const headingText = normalize(
-                        heading.innerText || heading.textContent || ''
-                    );
-                    if (headingText) return headingText;
-                }
-
-                const main = document.querySelector('main');
-                if (!main) return '';
-                const lines = (main.innerText || '')
-                    .split('\\n')
-                    .map(normalize)
-                    .filter(Boolean);
-                return lines[0] || '';
-            }"""
-        )
-        if not isinstance(display_name, str):
-            return None
-        display_name = display_name.strip()
-        return display_name or None
 
     async def _wait_for_message_surface(
         self, target: _ProfileMessageTarget
@@ -2904,7 +2552,7 @@ class LinkedInExtractor:
             logger.debug("Profile page did not load for %s", linkedin_username)
 
         await handle_modal_close(self._page)
-        display_name = await self._read_profile_display_name()
+        display_name = await self._profile_page._read_profile_display_name()
         if not display_name:
             raise LinkedInScraperException(
                 f"Could not resolve a display name for {linkedin_username}."
@@ -2960,7 +2608,7 @@ class LinkedInExtractor:
         try:
             for i, (section_name, suffix, is_overlay) in enumerate(requested_ordered):
                 if i > 0:
-                    await asyncio.sleep(_NAV_DELAY)
+                    await asyncio.sleep(NAV_DELAY)
 
                 url = base_url + suffix
                 try:
@@ -3407,22 +3055,22 @@ class LinkedInExtractor:
                 break
 
             elapsed = time.monotonic() - started
-            if page_num > 0 and elapsed + _NAV_DELAY + slowest_page > budget:
+            if page_num > 0 and elapsed + NAV_DELAY + slowest_page > budget:
                 logger.debug(
                     "Stopping after %d pages: %.1fs spent, another page costs "
                     "up to %.1fs and the budget is %.1fs",
                     page_num,
                     elapsed,
-                    _NAV_DELAY + slowest_page,
+                    NAV_DELAY + slowest_page,
                     budget,
                 )
                 break
 
             if page_num > 0:
-                await asyncio.sleep(_NAV_DELAY)
+                await asyncio.sleep(NAV_DELAY)
 
             # Started after the delay, because the prediction above adds
-            # `_NAV_DELAY` to `slowest_page` itself. Timing from before the
+            # `NAV_DELAY` to `slowest_page` itself. Timing from before the
             # sleep folds it into every page after the first and then charges
             # it a second time, which stops a page early for every two seconds
             # of delay the run has already paid for.
@@ -3852,7 +3500,7 @@ class LinkedInExtractor:
                 break
 
             if page_num > 0:
-                await asyncio.sleep(_NAV_DELAY)
+                await asyncio.sleep(NAV_DELAY)
 
             url = (
                 base_url
@@ -4010,57 +3658,13 @@ class LinkedInExtractor:
         network: list[str] | None = None,
         current_company: str | None = None,
     ) -> dict[str, Any]:
-        """Search for people and extract the results page.
-
-        Args:
-            keywords: Free-text query ("software engineer", "recruiter at Google").
-            location: Optional location filter ("New York", "Remote").
-            network: Optional connection-degree filter. Each element is one of
-                ``"F"`` (1st-degree), ``"S"`` (2nd-degree), ``"O"`` (3rd-degree
-                and beyond). Example: ``["F"]`` to only return 1st-degree
-                connections. Invalid tokens raise ``ValueError``.
-            current_company: Optional current-employer filter. LinkedIn's
-                ``currentCompany`` facet only filters on the numeric company
-                URN id (e.g. ``"1115"`` for SAP); plain company names are
-                accepted by the URL but ignored by LinkedIn and return the
-                unfiltered result set. Look up a company's URN via
-                ``get_company_profile`` -- it is exposed under
-                ``references["about"]``.
-
-        Returns:
-            {url, sections: {name: text}}
-        """
-        # Builds before it navigates, and the builder refuses a filter
-        # LinkedIn would swallow, so an invalid token costs no page load.
-        url = build_people_search_url(
+        """Search for people and extract the results page."""
+        return await self._person.search_people(
             keywords,
             location=location,
             network=network,
             current_company=current_company,
         )
-        extracted = await self.extract_page(url, section_name="search_results")
-
-        sections: dict[str, str] = {}
-        references: dict[str, list[Reference]] = {}
-        section_errors: dict[str, dict[str, Any]] = {}
-        if extracted.text and extracted.text != RATE_LIMITED_SECTION_TEXT:
-            sections["search_results"] = extracted.text
-            if extracted.references:
-                references["search_results"] = extracted.references
-        elif extracted.text == RATE_LIMITED_SECTION_TEXT:
-            section_errors["search_results"] = rate_limited_section_error()
-        elif extracted.error:
-            section_errors["search_results"] = extracted.error
-
-        result: dict[str, Any] = {
-            "url": url,
-            "sections": sections,
-        }
-        if references:
-            result["references"] = references
-        if section_errors:
-            result["section_errors"] = section_errors
-        return result
 
     async def search_companies(
         self,
